+++ elfpy/markets/hyperdrive/hyperdrive_actions.py
@@ -15,8 +15,7 @@ import elfpy.time as time
 import elfpy.types as types
 from elfpy.time.time import StretchedTimeFP
 from elfpy.utils.math import FixedPoint, FixedPointMath
-from elfpy.utils.math.update_weighted_average import (
-    update_weighted_average, update_weighted_average_fp)
+from elfpy.utils.math.update_weighted_average import update_weighted_average, update_weighted_average_fp
 
 if TYPE_CHECKING:
     import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
@@ -69,14 +68,6 @@ class MarketDeltas(base_market.MarketDeltas):
     short_checkpoints: defaultdict[float, float] = field(default_factory=lambda: defaultdict(float))
     total_supply_longs: defaultdict[float, float] = field(default_factory=lambda: defaultdict(float))
     total_supply_shorts: defaultdict[float, float] = field(default_factory=lambda: defaultdict(float))
-
-
-@types.freezable(frozen=True, no_new_attribs=True)
-@dataclass
-class MarketActionResult(base_market.MarketActionResult):
-    r"""The result to a market of performing a trade"""
-    d_base: float
-    d_bonds: float
 
 
 @types.freezable(frozen=False, no_new_attribs=True)
@@ -1096,14 +1087,6 @@ class MarketDeltasFP(base_market.MarketDeltas):
     total_supply_shorts: defaultdict[FixedPoint, FixedPoint] = field(
         default_factory=lambda: defaultdict(lambda: FixedPoint(0))
     )
-
-
-@types.freezable(frozen=True, no_new_attribs=True)
-@dataclass
-class MarketActionResultFP(base_market.MarketActionResultFP):
-    r"""The result to a market of performing a trade"""
-    d_base: FixedPoint
-    d_bonds: FixedPoint
 
 
 @types.freezable(frozen=False, no_new_attribs=True)
