+++ elfpy/pricing_models/hyperdrive.py
@@ -4,8 +4,8 @@ from __future__ import annotations  # types will be strings by default in 3.11
 from decimal import Decimal
 from typing import TYPE_CHECKING
 
-import elfpy.agents.agent as agent
-import elfpy.markets.hyperdrive.hyperdrive_actions as hyperdrive_actions
+import elfpy.agents.agent_trade_result as agent_trade_result
+import elfpy.markets.hyperdrive.market_action_result as market_action_result
 import elfpy.pricing_models.trades as trades
 import elfpy.pricing_models.yieldspace as yieldspace_pm
 import elfpy.time as time
@@ -144,20 +144,20 @@ class HyperdrivePricingModel(yieldspace_pm.YieldspacePricingModel):
         )
         # Compute the user's trade result including both the flat and the curve parts of the trade.
         if out.unit == types.TokenType.BASE:
-            user_result = agent.AgentTradeResult(
+            user_result = agent_trade_result.AgentTradeResult(
                 d_base=out.amount,
                 d_bonds=float(-flat_with_fee + Decimal(curve.user_result.d_bonds)),
             )
-            market_result = hyperdrive_actions.MarketActionResult(
+            market_result = market_action_result.MarketActionResult(
                 d_base=-out.amount,
                 d_bonds=curve.market_result.d_bonds,
             )
         elif out.unit == types.TokenType.PT:
-            user_result = agent.AgentTradeResult(
+            user_result = agent_trade_result.AgentTradeResult(
                 d_base=float(-flat_with_fee + Decimal(curve.user_result.d_base)),
                 d_bonds=out.amount,
             )
-            market_result = hyperdrive_actions.MarketActionResult(
+            market_result = market_action_result.MarketActionResult(
                 d_base=float(flat_with_fee + Decimal(curve.market_result.d_base)),
                 d_bonds=curve.market_result.d_bonds,
             )
@@ -291,20 +291,20 @@ class HyperdrivePricingModel(yieldspace_pm.YieldspacePricingModel):
         )
         # Compute the user's trade result including both the flat and the curve parts of the trade.
         if in_.unit == types.TokenType.BASE:
-            user_result = agent.AgentTradeResult(
+            user_result = agent_trade_result.AgentTradeResult(
                 d_base=-in_.amount,
                 d_bonds=float(flat_with_fee + Decimal(curve.user_result.d_bonds)),
             )
-            market_result = hyperdrive_actions.MarketActionResult(
+            market_result = market_action_result.MarketActionResult(
                 d_base=in_.amount,
                 d_bonds=curve.market_result.d_bonds,
             )
         elif in_.unit == types.TokenType.PT:
-            user_result = agent.AgentTradeResult(
+            user_result = agent_trade_result.AgentTradeResult(
                 d_base=float(flat_with_fee + Decimal(curve.user_result.d_base)),
                 d_bonds=-in_.amount,
             )
-            market_result = hyperdrive_actions.MarketActionResult(
+            market_result = market_action_result.MarketActionResult(
                 d_base=float(-flat_with_fee + Decimal(curve.market_result.d_base)),
                 d_bonds=curve.market_result.d_bonds,
             )
@@ -485,20 +485,20 @@ class HyperdrivePricingModelFP(yieldspace_pm.YieldspacePricingModelFP):
         )
         # Compute the user's trade result including both the flat and the curve parts of the trade.
         if out.unit == types.TokenType.BASE:
-            user_result = agent.AgentTradeResultFP(
+            user_result = agent_trade_result.AgentTradeResultFP(
                 d_base=out.amount,
                 d_bonds=-flat_with_fee + curve.user_result.d_bonds,
             )
-            market_result = hyperdrive_actions.MarketActionResultFP(
+            market_result = market_action_result.MarketActionResultFP(
                 d_base=-out.amount,
                 d_bonds=curve.market_result.d_bonds,
             )
         elif out.unit == types.TokenType.PT:
-            user_result = agent.AgentTradeResultFP(
+            user_result = agent_trade_result.AgentTradeResultFP(
                 d_base=-flat_with_fee + curve.user_result.d_base,
                 d_bonds=out.amount,
             )
-            market_result = hyperdrive_actions.MarketActionResultFP(
+            market_result = market_action_result.MarketActionResultFP(
                 d_base=flat_with_fee + curve.market_result.d_base,
                 d_bonds=curve.market_result.d_bonds,
             )
@@ -630,20 +630,20 @@ class HyperdrivePricingModelFP(yieldspace_pm.YieldspacePricingModelFP):
         )
         # Compute the user's trade result including both the flat and the curve parts of the trade.
         if in_.unit == types.TokenType.BASE:
-            user_result = agent.AgentTradeResultFP(
+            user_result = agent_trade_result.AgentTradeResultFP(
                 d_base=-in_.amount,
                 d_bonds=flat_with_fee + curve.user_result.d_bonds,
             )
-            market_result = hyperdrive_actions.MarketActionResultFP(
+            market_result = market_action_result.MarketActionResultFP(
                 d_base=in_.amount,
                 d_bonds=curve.market_result.d_bonds,
             )
         elif in_.unit == types.TokenType.PT:
-            user_result = agent.AgentTradeResultFP(
+            user_result = agent_trade_result.AgentTradeResultFP(
                 d_base=flat_with_fee + curve.user_result.d_base,
                 d_bonds=-in_.amount,
             )
-            market_result = hyperdrive_actions.MarketActionResultFP(
+            market_result = market_action_result.MarketActionResultFP(
                 d_base=-flat_with_fee + curve.market_result.d_base,
                 d_bonds=curve.market_result.d_bonds,
             )
